+++ ansible/roles/amazon-aws/amazon-aws/plugins/modules/cloudformation_facts.py
@@ -1,1 +1,5 @@
+XSym
+0022
+b7b646f8caf3ef28a09f058e2f500311
 cloudformation_info.py
+